rename pserver to ps
